+++ packages/lint-suite/src/lib/angular-template.ts
@@ -26,6 +26,14 @@ export const angularTemplate = defineConfig([
       '@angular-eslint/template/prefer-control-flow': 'error',
       '@angular-eslint/template/prefer-self-closing-tags': 'error',
       '@angular-eslint/template/prefer-ngsrc': 'error',
+      '@angular-eslint/template/prefer-empty-for': 'error',
+      '@angular-eslint/template/prefer-built-in-pipes': 'error',
+      '@angular-eslint/template/no-call-expression': 'error',
+      '@angular-eslint/template/no-interpolation-in-attributes': 'error',
+      '@angular-eslint/template/no-empty-control-flow-blocks': 'error',
+      '@angular-eslint/template/no-nested-tags': 'error',
+      '@angular-eslint/template/use-track-by-function': 'error',
+      '@angular-eslint/template/label-has-associated-control': 'error',
       '@angular-eslint/template/attributes-order': [
         'error',
         {