+++ packages/lint-suite/src/lib/playwright.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'eslint/config';
 export const playwright = defineConfig([
   {
     ...playwrightEslint.configs['flat/recommended'],
-    files: ['**/*.test.ts', '**/*.test.js'],
+    files: ['**/*.e2e.ts', '**/*.e2e.js'],
     rules: {
       ...playwrightEslint.configs['flat/recommended'].rules
     }